test: improve test coverage for SDP parsing and extraction logic
Moved inline/adjacent tests for SDP parsing (`tests.rs`, `raop_tests.rs`) to a dedicated `src/protocol/sdp/tests/` module directory structure (`mod.rs` and `raop.rs`) following the project convention for large test sets. Added extensive unit testing to cover invalid inputs, missing fields, base64 decoding errors, and different codec detections, greatly improving code coverage and robustness of the `src/protocol/sdp` submodules.

Additionally, updated formatting and cleared existing clippy warnings.
