+++ src/protocol/sdp/tests/raop.rs
@@ -1,0 +1,270 @@
+use crate::protocol::sdp::parser::SdpParser;
+use crate::protocol::sdp::raop::*;
+use crate::receiver::session::AudioCodec;
+
+const SAMPLE_SDP: &str = r"v=0
+o=iTunes 3413821438 0 IN IP4 192.168.1.100
+s=iTunes
+c=IN IP4 192.168.1.1
+t=0 0
+m=audio 0 RTP/AVP 96
+a=rtpmap:96 AppleLossless
+a=fmtp:96 352 0 16 40 10 14 2 255 0 0 44100
+a=rsaaeskey:VGhpcyBpcyBhIHRlc3Qga2V5IHRoYXQgaXMgdXNlZCBmb3IgdGVzdGluZw==
+a=aesiv:MDEyMzQ1Njc4OWFiY2RlZg==
+a=min-latency:11025
+";
+
+const SIMPLE_SDP: &str = r"v=0
+o=- 0 0 IN IP4 127.0.0.1
+s=AirTunes
+t=0 0
+m=audio 0 RTP/AVP 96
+a=rtpmap:96 AppleLossless
+a=fmtp:96 352 0 16 40 10 14 2 255 0 0 44100
+";
+
+#[test]
+fn test_detect_codec_alac() {
+    let sdp = SdpParser::parse(SIMPLE_SDP).unwrap();
+    let audio = sdp.audio_media().unwrap();
+
+    let codec = detect_codec(audio).unwrap();
+    assert_eq!(codec, AudioCodec::Alac);
+}
+
+#[test]
+fn test_parse_alac_parameters() {
+    let fmtp = "96 352 0 16 40 10 14 2 255 0 0 44100";
+    let params = AlacParameters::parse(fmtp).unwrap();
+
+    assert_eq!(params.frames_per_packet, 352);
+    assert_eq!(params.bit_depth, 16);
+    assert_eq!(params.channels, 2);
+    assert_eq!(params.sample_rate, 44100);
+}
+
+#[test]
+fn test_parse_alac_parameters_no_payload_type() {
+    let fmtp = "352 0 16 40 10 14 2 255 0 0 44100";
+    let params = AlacParameters::parse(fmtp).unwrap();
+
+    assert_eq!(params.frames_per_packet, 352);
+    assert_eq!(params.bit_depth, 16);
+    assert_eq!(params.channels, 2);
+    assert_eq!(params.sample_rate, 44100);
+}
+
+#[test]
+fn test_parse_alac_parameters_invalid_values() {
+    let fmtp = "96 352 0 16 40 10 14 2 255 0 0 invalid";
+    let result = AlacParameters::parse(fmtp);
+    assert!(result.is_err());
+}
+
+#[test]
+fn test_parse_encryption_params() {
+    let sdp = SdpParser::parse(SAMPLE_SDP).unwrap();
+    let audio = sdp.audio_media().unwrap();
+
+    let enc = parse_encryption(audio).unwrap();
+    assert!(enc.is_some());
+
+    let enc = enc.unwrap();
+    assert!(!enc.encrypted_aes_key.is_empty());
+    assert_eq!(enc.aes_iv.len(), 16);
+}
+
+#[test]
+fn test_no_encryption() {
+    let sdp = SdpParser::parse(SIMPLE_SDP).unwrap();
+    let audio = sdp.audio_media().unwrap();
+
+    let enc = parse_encryption(audio).unwrap();
+    assert!(enc.is_none());
+}
+
+#[test]
+fn test_extract_stream_params_unencrypted() {
+    let sdp = SdpParser::parse(SIMPLE_SDP).unwrap();
+
+    let params = extract_stream_parameters(&sdp, None).unwrap();
+
+    assert_eq!(params.codec, AudioCodec::Alac);
+    assert_eq!(params.sample_rate, 44100);
+    assert_eq!(params.bits_per_sample, 16);
+    assert_eq!(params.channels, 2);
+    assert_eq!(params.frames_per_packet, 352);
+    assert!(params.aes_key.is_none());
+}
+
+#[test]
+fn test_pcm_codec() {
+    let sdp_str = r"v=0
+o=- 0 0 IN IP4 127.0.0.1
+s=Test
+t=0 0
+m=audio 0 RTP/AVP 96
+a=rtpmap:96 L16/44100/2
+";
+    let sdp = SdpParser::parse(sdp_str).unwrap();
+    let audio = sdp.audio_media().unwrap();
+
+    let codec = detect_codec(audio).unwrap();
+    assert_eq!(codec, AudioCodec::Pcm);
+}
+
+#[test]
+fn test_min_latency_extraction() {
+    let sdp = SdpParser::parse(SAMPLE_SDP).unwrap();
+    let params = extract_stream_parameters(&sdp, None).unwrap();
+
+    assert_eq!(params.min_latency, Some(11025));
+}
+
+#[test]
+fn test_detect_codec_aac_eld() {
+    let sdp_str = r"v=0
+m=audio 0 RTP/AVP 96
+a=rtpmap:96 mpeg4-generic ELD
+";
+    let sdp = SdpParser::parse(sdp_str).unwrap();
+    let audio = sdp.audio_media().unwrap();
+    let codec = detect_codec(audio).unwrap();
+    assert_eq!(codec, AudioCodec::AacEld);
+}
+
+#[test]
+fn test_detect_codec_aac_lc() {
+    let sdp_str = r"v=0
+m=audio 0 RTP/AVP 96
+a=rtpmap:96 mpeg4-generic
+";
+    let sdp = SdpParser::parse(sdp_str).unwrap();
+    let audio = sdp.audio_media().unwrap();
+    let codec = detect_codec(audio).unwrap();
+    assert_eq!(codec, AudioCodec::AacLc);
+}
+
+#[test]
+fn test_detect_codec_mp4a_latm() {
+    let sdp_str = r"v=0
+m=audio 0 RTP/AVP 96
+a=rtpmap:96 MP4A-LATM
+";
+    let sdp = SdpParser::parse(sdp_str).unwrap();
+    let audio = sdp.audio_media().unwrap();
+    let codec = detect_codec(audio).unwrap();
+    assert_eq!(codec, AudioCodec::AacLc);
+}
+
+#[test]
+fn test_detect_codec_unknown() {
+    let sdp_str = r"v=0
+m=audio 0 RTP/AVP 96
+a=rtpmap:96 UnknownCodec
+";
+    let sdp = SdpParser::parse(sdp_str).unwrap();
+    let audio = sdp.audio_media().unwrap();
+    let codec = detect_codec(audio);
+    assert_eq!(codec, None);
+}
+
+#[test]
+fn test_parse_encryption_invalid_aesiv_base64() {
+    let sdp_str = r"v=0
+m=audio 0 RTP/AVP 96
+a=rsaaeskey:VGhpcyBpcyBhIHRlc3Qga2V5IHRoYXQgaXMgdXNlZCBmb3IgdGVzdGluZw==
+a=aesiv:!!invalid!!base64!!
+";
+    let sdp = SdpParser::parse(sdp_str).unwrap();
+    let audio = sdp.audio_media().unwrap();
+    let res = parse_encryption(audio);
+    assert!(res.is_err());
+}
+
+#[test]
+fn test_parse_encryption_invalid_rsaaeskey_base64() {
+    let sdp_str = r"v=0
+m=audio 0 RTP/AVP 96
+a=rsaaeskey:!!invalid!!base64!!
+a=aesiv:MDEyMzQ1Njc4OWFiY2RlZg==
+";
+    let sdp = SdpParser::parse(sdp_str).unwrap();
+    let audio = sdp.audio_media().unwrap();
+    let res = parse_encryption(audio);
+    assert!(res.is_err());
+}
+
+#[test]
+fn test_parse_encryption_invalid_aesiv_length() {
+    let sdp_str = r"v=0
+m=audio 0 RTP/AVP 96
+a=rsaaeskey:VGhpcyBpcyBhIHRlc3Qga2V5IHRoYXQgaXMgdXNlZCBmb3IgdGVzdGluZw==
+a=aesiv:MDEyMzQ1Njc4OQ==
+"; // Shorter than 16 bytes
+    let sdp = SdpParser::parse(sdp_str).unwrap();
+    let audio = sdp.audio_media().unwrap();
+    let res = parse_encryption(audio);
+    assert!(res.is_err());
+}
+
+#[test]
+fn test_extract_stream_parameters_missing_audio_media() {
+    let sdp_str = r"v=0
+m=video 0 RTP/AVP 96
+a=rtpmap:96 H264
+";
+    let sdp = SdpParser::parse(sdp_str).unwrap();
+    let res = extract_stream_parameters(&sdp, None);
+    assert!(res.is_err());
+}
+
+#[test]
+fn test_extract_stream_parameters_missing_rtpmap() {
+    let sdp_str = r"v=0
+m=audio 0 RTP/AVP 96
+a=fmtp:96 352 0 16 40 10 14 2 255 0 0 44100
+";
+    let sdp = SdpParser::parse(sdp_str).unwrap();
+    let res = extract_stream_parameters(&sdp, None);
+    assert!(res.is_err());
+}
+
+#[test]
+fn test_extract_stream_parameters_alac_missing_fmtp() {
+    let sdp_str = r"v=0
+m=audio 0 RTP/AVP 96
+a=rtpmap:96 AppleLossless
+";
+    let sdp = SdpParser::parse(sdp_str).unwrap();
+    let res = extract_stream_parameters(&sdp, None);
+    assert!(res.is_err());
+}
+
+#[test]
+fn test_extract_stream_parameters_aac() {
+    let sdp_str = r"v=0
+m=audio 0 RTP/AVP 96
+a=rtpmap:96 mpeg4-generic
+";
+    let sdp = SdpParser::parse(sdp_str).unwrap();
+    let params = extract_stream_parameters(&sdp, None).unwrap();
+    assert_eq!(params.codec, AudioCodec::AacLc);
+    assert_eq!(params.sample_rate, 44100);
+    assert_eq!(params.bits_per_sample, 16);
+    assert_eq!(params.channels, 2);
+    assert_eq!(params.frames_per_packet, 352);
+}
+
+#[test]
+fn test_parse_encryption_missing_aesiv() {
+    let sdp_str = r"v=0
+m=audio 0 RTP/AVP 96
+a=rsaaeskey:VGhpcyBpcyBhIHRlc3Qga2V5IHRoYXQgaXMgdXNlZCBmb3IgdGVzdGluZw==
+";
+    let sdp = SdpParser::parse(sdp_str).unwrap();
+    let audio = sdp.audio_media().unwrap();
+    let res = parse_encryption(audio);
+    assert!(res.is_err());
+}
